requirements.txt
Flask
python-dotenv
binance-futures-connector
